+++ lib/view/components/events/event_card.dart
