+++ lib/view/components/events/joined_event_card.dart
